documented g2o graph optimizer

+++ src/mapping_localization/src/publisher/cloud_publisher.cpp
@@ -3,7 +3,7 @@
  * @Created Date: 2020-02-05 02:27:30
  * @Author: Ren Qian
  * -----
- * @Last Modified: 2021-11-25 18:53:18
+ * @Last Modified: 2021-12-12 16:06:46
  * @Modified By: Xiaotao Guo
  */
 
@@ -12,17 +12,13 @@
 #include "glog/logging.h"
 
 namespace mapping_localization {
-CloudPublisher::CloudPublisher(ros::NodeHandle& nh,
-                               std::string topic_name,
-                               std::string frame_id,
-                               size_t buff_size)
+CloudPublisher::CloudPublisher(ros::NodeHandle& nh, std::string topic_name, std::string frame_id, size_t buff_size)
     : nh_(nh), frame_id_(frame_id) {
     publisher_ = nh_.advertise<sensor_msgs::PointCloud2>(topic_name, buff_size);
 }
 
-void CloudPublisher::Publish(CloudData::Cloud_Ptr& cloud_ptr_input,
-                             double time) {
-    ros::Time ros_time((float)time);
+void CloudPublisher::Publish(CloudData::Cloud_Ptr& cloud_ptr_input, double time) {
+    ros::Time ros_time(time);
     PublishData(cloud_ptr_input, ros_time);
 }
 
@@ -31,10 +27,8 @@ void CloudPublisher::Publish(CloudData::Cloud_Ptr& cloud_ptr_input) {
     PublishData(cloud_ptr_input, time);
 }
 
-void CloudPublisher::PublishData(CloudData::Cloud_Ptr& cloud_ptr_input,
-                                 ros::Time time) {
-    sensor_msgs::PointCloud2Ptr cloud_ptr_output(
-        new sensor_msgs::PointCloud2());
+void CloudPublisher::PublishData(CloudData::Cloud_Ptr& cloud_ptr_input, ros::Time time) {
+    sensor_msgs::PointCloud2Ptr cloud_ptr_output(new sensor_msgs::PointCloud2());
     pcl::toROSMsg(*cloud_ptr_input, *cloud_ptr_output);
 
     cloud_ptr_output->header.stamp = time;
@@ -42,7 +36,5 @@ void CloudPublisher::PublishData(CloudData::Cloud_Ptr& cloud_ptr_input,
     publisher_.publish(*cloud_ptr_output);
 }
 
-bool CloudPublisher::HasSubscribers() {
-    return publisher_.getNumSubscribers() != 0;
-}
+bool CloudPublisher::HasSubscribers() { return publisher_.getNumSubscribers() != 0; }
 }  // namespace mapping_localization
