+++ src/mapping_localization/src/publisher/key_frame_publisher.cpp
@@ -3,7 +3,7 @@
  * @Created Date: 2020-02-06 21:11:44
  * @Author: Ren Qian
  * -----
- * @Last Modified: 2021-11-24 00:38:07
+ * @Last Modified: 2021-12-12 16:06:53
  * @Modified By: Xiaotao Guo
  */
 
@@ -12,19 +12,15 @@
 #include <Eigen/Dense>
 
 namespace mapping_localization {
-KeyFramePublisher::KeyFramePublisher(ros::NodeHandle& nh,
-                                     std::string topic_name,
-                                     std::string frame_id,
-                                     int buff_size)
+KeyFramePublisher::KeyFramePublisher(ros::NodeHandle& nh, std::string topic_name, std::string frame_id, int buff_size)
     : nh_(nh), frame_id_(frame_id) {
-    publisher_ =
-        nh_.advertise<geometry_msgs::PoseWithCovarianceStamped>(topic_name, buff_size);
+    publisher_ = nh_.advertise<geometry_msgs::PoseWithCovarianceStamped>(topic_name, buff_size);
 }
 
 void KeyFramePublisher::Publish(KeyFrame& key_frame) {
     geometry_msgs::PoseWithCovarianceStamped pose_stamped;
 
-    ros::Time ros_time((float)key_frame.time);
+    ros::Time ros_time(key_frame.time);
     pose_stamped.header.stamp = ros_time;
     pose_stamped.header.frame_id = frame_id_;
